fix(command): Do not use `settings.TESTING` anymore
`settings.TESTING` was used to disable `stdout` and `stderr` during testing,
but it is not a standard Django settings. `stdout` and `stderr` must instead be
redirected using their correpsonding argument of `call_command`[1].

If the commands are not being directly tested, it is also better to use
`Document` and `Document._index` to manage indices and documents.

[1] https://docs.djangoproject.com/en/4.0/topics/testing/tools/#management-commands

Issues: #10