feat(indices): manage Index versions
An Index version is a concrete OpenSearch index that is named after a django_opensearch_dsl Index but with a timestamp suffix.

Making an Index version active is nothing more than creating an Index alias on this version to the django_opensearch_dsl Index.

This will allow to have multiple concrete indices living for a given Document Index, making it possible to perform zero-downtime full reindexing.

This commit adds basic management of these Index versions:
* Listing versions for a given Index
* Getting the active version
* Creating a new version
* Activating a version by swapping an alias (removing the alias for the active version and setting the alias on the given one)